Plot posterior distributions from a fahb analysis

Takes an object of class fahb_analysis and plots the posterior distributions of the predicted time for the trial to recruit and of the three model parameters.

Usage

# S3 method for class 'fahb_analysis'
plot(x, ...)

Arguments

x

object of class fahb_analysis as produced by fahb_analysis().

...

further arguments passed to or from other methods.

Value

no return value, called for side effects.